Move enemy health bar to separate file
