+++ src/main/java/me/bobthe28th/bday/games/ctf/CTF.java
@@ -2,7 +2,7 @@ package me.bobthe28th.bday.games.ctf;
 
 import me.bobthe28th.bday.Main;
 import me.bobthe28th.bday.games.Game;
-import me.bobthe28th.bday.games.GamePlayer;
+import me.bobthe28th.bday.games.player.GamePlayer;
 import me.bobthe28th.bday.games.GameState;
 import me.bobthe28th.bday.scoreboard.ScoreboardObjective;
 import org.bukkit.ChatColor;
