+++ src/main/java/me/bobthe28th/bday/games/ctf/CTFPlayer.java
@@ -1,0 +1,7 @@
+package me.bobthe28th.bday.games.ctf;
+
+public abstract class CTFPlayer {
+
+    abstract void spawn();
+    
+}
