+++ src/main/java/me/bobthe28th/bday/games/player/GameModePlayer.java
@@ -1,0 +1,4 @@
+package me.bobthe28th.bday.games.player;
+
+public class GameModePlayer {
+}
