+++ src/main/java/me/bobthe28th/bday/games/player/GamePlayer.java
@@ -1,0 +1,62 @@
+package me.bobthe28th.bday.games.player;
+
+import me.bobthe28th.bday.Main;
+import me.bobthe28th.bday.scoreboard.ScoreboardController;
+import org.bukkit.entity.LivingEntity;
+import org.bukkit.entity.Player;
+import org.bukkit.potion.PotionEffect;
+
+public class GamePlayer {
+
+    protected final Player player;
+    protected final Main plugin;
+    private final ScoreboardController scoreboardController;
+    private final EnemyHealthBar enemyHealthBar;
+    boolean isAlive;
+
+    public GamePlayer(Main plugin, Player player) {
+        //plugin.getServer().getPluginManager().registerEvents(this, plugin);
+        this.player = player;
+        this.plugin = plugin;
+        this.player.setLevel(0);
+        this.player.setExp(0.0F);
+        this.player.setFoodLevel(20);
+        this.player.setSaturation(0F);
+        this.player.setGlowing(false);
+        this.player.setInvisible(false);
+        this.player.setPlayerListHeaderFooter("Deez", "Nuts");
+        for (PotionEffect effect : player.getActivePotionEffects()) {
+            player.removePotionEffect(effect.getType());
+        }
+        scoreboardController = new ScoreboardController(this);
+        enemyHealthBar = new EnemyHealthBar(this, plugin);
+    }
+
+    public Player getPlayer() {
+        return player;
+    }
+
+    public ScoreboardController getScoreboardController() {
+        return scoreboardController;
+    }
+
+
+    public void damage(LivingEntity entity, double damage) {
+
+        enemyHealthBar.startEnemyHealthCooldown();
+        enemyHealthBar.updateEnemyHealth(entity, damage);
+    }
+
+    public EnemyHealthBar getEnemyHealthBar() {
+        return enemyHealthBar;
+    }
+
+    public void remove() {
+        removeNotMap();
+        plugin.getGameManager().getGamePlayers().remove(player);
+    }
+
+    public void removeNotMap() {
+        //HandlerList.unregisterAll(this);
+    }
+}
